ui界面 1 期 准备工作，导入文件，改成vs2022

+++ LurenjiaEngine/LurenjiaEngine/simple_library/private/simple_math/simple_math.c
@@ -1,4 +1,4 @@
-#include "../../public/simple_math/simple_math.h"
+#include "simple_library/public/simple_math/simple_math.h"
 
 #ifndef RAND_MAX
 #define RAND_MAX 0x7fff
@@ -8,7 +8,7 @@ unsigned int get_uint32_random(unsigned int in_max)
 {
 	if (in_max > 0)
 	{
-		return rand() / (double)RAND_MAX * (in_max - 0) + 0;
+		return (unsigned int)(rand() / (double)RAND_MAX * (in_max - 0) + 0);
 	}
 
 	return 0;
